Applied aggregate pattern so that all plots and visualizations come under the same module for better clarity and order
