+++ src/covidify/data_prep.py
@@ -24,7 +24,7 @@ from datetime import datetime, date, time
 from covidify.sources import github, wiki
 from covidify.config import REPO, TMP_FOLDER, TMP_GIT, DATA
 from covidify.utils.utils import replace_arg_score
-from covidify.src.covidify.output_destination import output_data_folder
+from covidify.output_destination import output_data_folder
 
 
 args = docopt.docopt(__doc__)
