+++ src/covidify/forecast.py
@@ -25,7 +25,8 @@ from dateutil.parser import parse
 from pmdarima.arima import auto_arima
 from datetime import datetime, date, time 
 from sklearn.metrics import mean_squared_error
-from covidify.config import PERC_SPLIT, FIG_SIZE
+from covidify.config import PERC_SPLIT
+from covidify.data_visualization import plot_forecast
 
 font = {'weight' : 'bold',
         'size'   : 22}
@@ -60,35 +61,6 @@ if not os.path.exists(image_dir):
     print('Creating reports folder...')
     os.system('mkdir -p ' + image_dir)
 
-
-def plot_forecast(tmp_df, train, index_forecast, forecast, confint):
-    '''
-    Plot the values of train and test, the predictions from ARIMA and the shadowing
-    for the confidence interval.
-    
-    '''
-
-    # For shadowing
-    lower_series = pd.Series(confint[:, 0], index=index_forecast)
-    upper_series = pd.Series(confint[:, 1], index=index_forecast)
-    
-    print('... saving graph')
-    fig, ax = plt.subplots(figsize=FIG_SIZE)
-    plt.title('ARIMA - Prediction for cumalitive case counts {} days in the future'.format(days_in_future))    
-    plt.plot(tmp_df.cumulative_cases, label='Train',marker='o')
-    plt.plot(tmp_df.pred, label='Forecast', marker='o')
-    tmp_df.groupby('date')[['']].sum().plot(ax=ax)
-    plt.fill_between(index_forecast, 
-                     upper_series, 
-                     lower_series, 
-                     color='k', alpha=.1)
-    plt.ylabel('Infections')
-    plt.xlabel('Date')
-    fig.legend().set_visible(True)
-    fig = ax.get_figure()
-    fig.savefig(os.path.join(image_dir, 'cumulative_forecasts.png'))
-
-
 def forecast(tmp_df, train, index_forecast, days_in_future):
     
     # Fit model with training data
@@ -107,7 +79,7 @@ def forecast(tmp_df, train, index_forecast, days_in_future):
     forecast_df.to_csv(os.path.join(data_dir, forecast_file))
         
     plot_forecast(forecast_df, train, index_forecast, forecast, confint)
-    
+
 if __name__ == '__main__':
     print('Training forecasting model...')
